chore: migrate some components from features to components
