+++ apps/mobile/src/components/ColorScales/demos/index.ts
@@ -1,0 +1,3 @@
+export { default as BasicDemo } from './basic';
+export { default as FullDemo } from './full';
+export { default as TokenDemo } from './token';
