+++ apps/mobile/src/features/playground/components/toast.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ComponentPlayground, { DemoItem } from '../Playground';
+import ComponentPlayground, { DemoItem } from '../../../components/Playground';
 import {
   AdvancedDemo,
   BasicDemo,
